Bug fix, cleanup
