+++ c2cpg-tests/src/test/scala/io/shiftleft/c2cpg/querying/CfgTests.scala
@@ -1,9 +1,6 @@
 package io.shiftleft.c2cpg.querying
 
-import io.shiftleft.c2cpg.testfixtures.{CCodeToCpgSuite, DataFlowCodeToCpgSuite}
-import io.shiftleft.codepropertygraph.generated.DispatchTypes
-import io.shiftleft.codepropertygraph.generated.nodes.Call
-import io.shiftleft.dataflowengineoss.language.toExtendedCfgNode
+import io.shiftleft.c2cpg.testfixtures.CCodeToCpgSuite
 import io.shiftleft.semanticcpg.language._
 
 class CfgTests extends CCodeToCpgSuite {
@@ -45,43 +42,5 @@ class CfgTests extends CCodeToCpgSuite {
 
   "should find that method does not post dominate anything" in {
     cpg.method("foo").postDominates.l.size shouldBe 0
-  }
-}
-
-class CfgMacroTests extends DataFlowCodeToCpgSuite {
-  override val code: String =
-    """
-       #define MP4_GET4BYTES( dst ) MP4_GETX_PRIVATE( dst, GetDWBE(p_peek), 4 )
-       #define MP4_GETX_PRIVATE(dst, code, size) \
-    do \
-    { \
-        if( (i_read) >= (size) ) \
-        { \
-            dst = (code); \
-            p_peek += (size); \
-            i_read -= (size); \
-        } \
-        else \
-        { \
-            dst = 0; \
-            i_read = 0; \
-        } \
-    } while(0)
-
-    int foo() {
-       unsigned int x;
-       MP4_GET4BYTES(x);
-       sink(x);
-    }
-
-    """.stripMargin
-
-  "should create correct CFG for macro expansion and find data flow" in {
-    val List(callToMacro: Call) = cpg.method("foo").call.dispatchType(DispatchTypes.INLINED).l
-    callToMacro.argument.code.l shouldBe List("x")
-    callToMacro.cfgNext.code.toSet shouldBe Set("x", "i_read")
-    val source = cpg.method("foo").call.name("MP4_GET4BYTES").argument(1).l
-    val sink = cpg.method("foo").call.name("sink").argument(1).l
-    sink.reachableByFlows(source).l.foreach(println)
   }
 }
